+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to:
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ***
+
+## [1.8.0] - 2019-07-08 - RIHY
+
+### Added
+
+* `plot_stats()` method added to `Multiple()` class in `dyn_analysis` module. 
+  This can be used to present results from multiple time-stepping analysese. 
+  Refer Train Analysis example script for demo of its usage.
+  
+### Changed
+
+* `collate_stats()` method of `Multiple()` class in `dyn_analysis` module now 
+  returns a Pandas DataFrame, with MultiIndex columns and index to organise 
+  data hierachically
+  
+* `dashpot` argument added to `TMD().__init___()` method, to allow dashpot of 
+  TMDs to be defined (rather than TMD damping ratio) 
+  
+* Inputs used to define TMDs are now recorded as class attributes, to allow 
+  later retrieval (but not editing; no set methods are provided)
+  
+* Tutorial updated, with HTML export file included for accessibility
+
 ## [1.7.5] - 2019-05-14 - RIHY
 
 ### Fixed